pyproject.toml
[project]
name = "svip"
version = "1.0.0"
description = "SVIP v1.0 — 慢变量投资池系统（Slow Variable Investment Pool）"
authors = [
    {name = "SVIP Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
auto = [
    "yfinance>=0.2.30",
    "fredapi>=0.5.0",
    "akshare>=1.14.0",
    "requests>=2.31.0",
]
dashboard = [
    "streamlit>=1.38.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]
all = [
    "svip[auto,dashboard,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
